Huge structural changes MASSIVE overhaul
Made the whole program way more object oriented, it began as more or less a script and now it is a real boy program. With a separated Player class, Game State, and CoreGamplayLoop. All to get the victory condition to work, now it does and I realize why too. Learned so Much.